* Removed uk from Test-lang for 'cyrillic' task.
Due to this 'cyrillic' and 'cyrillic-desktop' tasks will not be automatically
installed during installation in Ukrainian. Some space will be saved because of
that.

  * Added character-cell unicode fonts from 'cyrillic-desktop' task to
    'ukrainian-desktop' task.
I am not sure whether anyone uses those fonts, but let's put them in
ukrainian-desktop to be safe.

That changes were tested during full (with desktop task) install in Ukrainian,
using version of tasksel with included package list.