Avoid installing kazakh-desktop by default when desktop task is selected.
A rebuild seems to have fixed this bug. Diffing the debian-tasks.desc,
it seems the kazakh task was just not included in the previous release:

+Task: kazakh
+Section: l10n
+Key:
+  task-kazakh
+Test-lang: kk
+
 Task: kazakh-desktop
 Section: l10n
 Key:
   task-kazakh-desktop
-Enhances: desktop
+Enhances: desktop, kazakh

This is odd, because both tasks above were added in the same commit and
have not been modified since. So I don't know what happened, unfortunatly.