switch default desktop task to xfce
This ensures that the desktop will fit on CD#1, which gnome currently does not.

There may be other reasons to prefer xfce as the default as well, but that
is a complex and subjective topic. Unfortunatly, Debian does not have a
well-defined procedure for making such choices, though it certianly has
well-defined procedures for reviewing them. So, I've decided to be bold,
and continue the tradition of making an arbitrary desktop selection for
Debian in tasksel.