Set APT::Acquire::Retries=3 to deal with minor network glitches. Thanks, Petter Reinholdtsen. Closes: #761970
Perhaps this would be a good default for apt in general? In any case, it
seems fine to do it in the installer.

Note that this could override an existing user configuration for this
value. If that turns out to be a problem, it could be done only when
in new install mode.