switch default desktop to Gnome
Based on prelimary results from
https://wiki.debian.org/DebianDesktop/Requalification/Jessie

Some desired data is not yet available, but at this point I'm around 80%
sure that gnome is coming out ahead in the process. This is particularly
based on accessibility and to some extent systemd integration.

Accessibility: Gnome and Mate are ahead by a large margin. Some of the
other desktops have had their accessability integration in Debian improved,
partly driven by this process, but still need significant upstream work.

Systemd/etc integration: Xfce, Mate, etc are stuck paying catch-up to
ongoing changes in this area. There will be time to hopefully iron these
issues out during the freeze once the tech stack stops changing out from
under them, so this is not a complete blocker for those desktops, but
going by the current status, Gnome is ahead.

The only single factor that I think could outweigh the above is media size,
if there was a strong desire by Debian to see a single CD with a standalone
usable desktop. However, the Debian live team doesn't care about fitting on
a traditional CD; and while the Debian CD team hasn't made a statement, my
impression as a member is that this is not something we care enough about
any more to make it a hard blocker on the default desktop.

Other less tangible things that influenced this decision slightly include:

- The Debian gnome team made a passionate case for Gnome having a bigger
  community, etc.
- Gnome 3 seems to have improved quite a lot since the last Debian release.
- The Debian XFCE team is ambivilant about whether it should be the default
  at all. They have not seen much added contributions from it being the
  default in testing for the past ~9 months and remain a small team.
- The Debian Mate team is making a pretty good case for Mate,
  but OTOH it's new in Debian, without much testing or many users.
  While at the same time being basically gnome 2.0. I am uncomfortable
  with Debian regressing, despite that being a fine desktop environment.
- Tasksel allows picking other desktops from a list, so this is really
  just the default, which can be easily changed.

(I will not be answering media enquiries related to this commit.)