Include MATE in the desktop task list, also based on desktop requalification results.
MATE has good accessibility, and is at least as good as XCFE in systemd
integration. And including it on the list provides a very familar option
for older Debian users.

Note that I have not tested MATE integration with d-i, particularly whether
d-i sets up graphical sudo for it when the system is installed with no root
password. That may need to be sorted out, and so this addition is only
preliminary.